requirements.txt
certifi==2022.12.7
dnspython==2.3.0
elastic-transport==8.4.0
elasticsearch==8.6.2
numpy==1.24.2
pandas==1.5.3
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
urllib3==1.26.14
